% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_tale_of_two_cities.R
\name{a_tale_of_two_cities}
\alias{a_tale_of_two_cities}
\title{Great Expectations}
\usage{
a_tale_of_two_cities()
}
\value{
tibble with 4 columns. "text", "book", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "Great Expectations".
}
\examples{
\dontrun{
a_tale_of_two_cities()
}
}
