% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adventures_of_huckleberry_finn.R
\name{adventures_of_huckleberry_finn}
\alias{adventures_of_huckleberry_finn}
\title{Adventures of Huckleberry Finn}
\usage{
adventures_of_huckleberry_finn()
}
\value{
tibble with 2 columns. "text" and "chapter".
}
\description{
Downlaod and lightly preprocess "Adventures of Huckleberry Finn".
}
\examples{
\dontrun{
adventures_of_huckleberry_finn()
}
}
