% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adventures_of_tom_sawyer.R
\name{adventures_of_tom_sawyer}
\alias{adventures_of_tom_sawyer}
\title{The Adventures of Tom Sawyer}
\usage{
adventures_of_tom_sawyer()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "The Adventures of Tom Sawyer".
}
\examples{
\dontrun{
adventures_of_tom_sawyer()
}
}
