% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aladdin_and_the_magic_lamp.R
\name{aladdin_and_the_magic_lamp}
\alias{aladdin_and_the_magic_lamp}
\title{Aladdin and the Magic Lamp
Downlaod and lightly preprocess "Aladdin and the Magic Lamp".}
\usage{
aladdin_and_the_magic_lamp()
}
\value{
tibble with 1 column. "text".
}
\description{
Aladdin and the Magic Lamp
Downlaod and lightly preprocess "Aladdin and the Magic Lamp".
}
\examples{
\dontrun{
aladdin_and_the_magic_lamp()
}
}
