% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alices_adventures.R
\name{alices_adventures}
\alias{alices_adventures}
\title{Alice's Adventures in Wonderland}
\usage{
alices_adventures()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "Alice's Adventures in Wonderland".
}
\examples{
\dontrun{
alices_adventures()
}
}
