% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anna_karenina.R
\name{anna_karenina}
\alias{anna_karenina}
\title{Anna Karenina}
\usage{
anna_karenina()
}
\value{
tibble with 2 columns. "text", "part" and "chapter".
}
\description{
Downlaod and lightly preprocess "Anna Karenina".
}
\examples{
\dontrun{
anna_karenina()
}
}
