% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beauty_and_the_beast.R
\name{beauty_and_the_beast}
\alias{beauty_and_the_beast}
\title{Beauty and the Beast}
\usage{
beauty_and_the_beast()
}
\value{
tibble with 1 column. "text".
}
\description{
Downlaod and lightly preprocess "Beauty and the Beast".
}
\examples{
\dontrun{
beauty_and_the_beast()
}
}
