% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_jekyll_and_mr_hyde.R
\name{dr_jekyll_and_mr_hyde}
\alias{dr_jekyll_and_mr_hyde}
\title{The Strange Case Of Dr. Jekyll And Mr. Hyde}
\usage{
dr_jekyll_and_mr_hyde()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "The Strange Case Of Dr. Jekyll And Mr.
Hyde".
}
\examples{
\dontrun{
dr_jekyll_and_mr_hyde()
}
}
