% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracula.R
\name{dracula}
\alias{dracula}
\title{Dracula}
\usage{
dracula()
}
\value{
tibble with 2 columns. "text" and "chapter".
}
\description{
Downlaod and lightly preprocess "Dracula".
}
\examples{
\dontrun{
dracula()
}
}
