% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankenstein.R
\name{frankenstein}
\alias{frankenstein}
\title{Frankenstein}
\usage{
frankenstein()
}
\value{
tibble with 2 columns. "text" and "chapter".
}
\description{
Downlaod and lightly preprocess "Frankenstein".
}
\examples{
\dontrun{
frankenstein()
}
}
