% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_expectations.R
\name{great_expectations}
\alias{great_expectations}
\title{Great Expectations}
\usage{
great_expectations()
}
\value{
tibble with 2 columns. "text" and "chapter".
}
\description{
Downlaod and lightly preprocess "Great Expectations".
}
\examples{
\dontrun{
great_expectations()
}
}
