% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jane_eyre.R
\name{jane_eyre}
\alias{jane_eyre}
\title{Jane Eyre}
\usage{
jane_eyre()
}
\value{
tibble with 3 columns. "text" and "chapter".
}
\description{
Downlaod and lightly preprocess "Jane Eyre".
}
\examples{
\dontrun{
jane_eyre()
}
}
