% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journey_to_the_centre_of_the_earth.R
\name{journey_to_the_centre_of_the_earth}
\alias{journey_to_the_centre_of_the_earth}
\title{A Journey to the Centre of the Earth}
\usage{
journey_to_the_centre_of_the_earth()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "A Journey to the Centre of the Earth".
}
\examples{
\dontrun{
journey_to_the_centre_of_the_earth()
}
}
