% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/king_arthur.R
\name{king_arthur}
\alias{king_arthur}
\title{The Legends Of King Arthur And His Knights}
\usage{
king_arthur()
}
\value{
tibble with 2 columns. "text" and "chapter".
}
\description{
Downlaod and lightly preprocess "The Legends Of King Arthur And His Knights".
}
\examples{
\dontrun{
king_arthur()
}
}
