% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leagues_under_the_sea.R
\name{leagues_under_the_sea}
\alias{leagues_under_the_sea}
\title{Twenty Thousand Leagues under the Sea}
\usage{
leagues_under_the_sea()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "Twenty Thousand Leagues under the Sea".
}
\examples{
\dontrun{
leagues_under_the_sea()
}
}
