% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_women.R
\name{little_women}
\alias{little_women}
\title{Little Women}
\usage{
little_women()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "Little Women".
}
\examples{
\dontrun{
little_women()
}

}
