% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middlemarch.R
\name{middlemarch}
\alias{middlemarch}
\title{Great Expectations}
\usage{
middlemarch()
}
\value{
tibble with 3 columns. "text", "book" and "chapter".
}
\description{
Downlaod and lightly preprocess "Great Expectations".
}
\examples{
\dontrun{
middlemarch()
}
}
