% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peter_pan.R
\name{peter_pan}
\alias{peter_pan}
\title{Peter Pan}
\usage{
peter_pan()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "Peter Pan".
}
\examples{
\dontrun{
peter_pan()
}
}
