% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantom_of_the_opera.R
\name{phantom_of_the_opera}
\alias{phantom_of_the_opera}
\title{The Phantom of the Opera}
\usage{
phantom_of_the_opera()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "The Phantom of the Opera".
}
\examples{
\dontrun{
phantom_of_the_opera()
}
}
