% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robin_hood.R
\name{robin_hood}
\alias{robin_hood}
\title{The Merry Adventures of Robin Hood}
\usage{
robin_hood()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "The Merry Adventures of Robin Hood".
}
\examples{
\dontrun{
robin_hood()
}
}
