% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_prophet.R
\name{the_prophet}
\alias{the_prophet}
\title{The Prophet}
\usage{
the_prophet()
}
\value{
tibble with 1 column. "text".
}
\description{
Downlaod and lightly preprocess "The Prophet".
}
\examples{
\dontrun{
the_prophet()
}
}
