% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_scarlet_letter.R
\name{the_scarlet_letter}
\alias{the_scarlet_letter}
\title{The Scarlet Letter}
\usage{
the_scarlet_letter()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "The Scarlet Letter".
}
\examples{
\dontrun{
the_scarlet_letter()
}
}
