% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treasure_island.R
\name{treasure_island}
\alias{treasure_island}
\title{Treasure Island}
\usage{
treasure_island()
}
\value{
tibble with 4 columns. "text", "chapter", "part" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "Treasure Island".
}
\examples{
\dontrun{
treasure_island()
}
}
