% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/war_and_peace.R
\name{war_and_peace}
\alias{war_and_peace}
\title{War and Peace}
\usage{
war_and_peace()
}
\value{
tibble with 3 columns. "text", "book" and "chapter".
}
\description{
Downlaod and lightly preprocess "War and Peace".
}
\examples{
\dontrun{
war_and_peace()
}
}
