% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizard_of_oz.R
\name{wizard_of_oz}
\alias{wizard_of_oz}
\title{The Wonderful Wizard of Oz}
\usage{
wizard_of_oz()
}
\value{
tibble with 3 columns. "text", "chapter" and "chapter_name".
}
\description{
Downlaod and lightly preprocess "The Wonderful Wizard of Oz".
}
\examples{
\dontrun{
wizard_of_oz()
}
}
