% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuthering_heights.R
\name{wuthering_heights}
\alias{wuthering_heights}
\title{Wuthering Heights}
\usage{
wuthering_heights()
}
\value{
tibble with 2 columns. "text" and "chapter".
}
\description{
Downlaod and lightly preprocess "Wuthering Heights".
}
\examples{
\dontrun{
wuthering_heights()
}
}
